ui: add generic behaviour for visible menus misc fixes and improvements: ui_grid: use window bg as default bg ui_image: white_is_transparent option ui_input: set_input_placeholder fix apps freezing reading 0 bytes fix api_get_window_position move apps btn a bit

+++ draw.c
@@ -9,7 +9,7 @@ inline uint16_t rgb16(uint8_t r, uint8_t g, uint8_t b) {
    return ((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3);
 }
 
-inline void setpixel_safe(surface_t *surface, int index, int colour) {
+inline void setpixel_safe(surface_t *surface, int index, uint16_t colour) {
    if(index < 0 || index >= surface->width*surface->height) {
       //window_writestr("Attempted to write outside framebuffer bounds\n", 0, 0);
    } else {
